data_collection_and_etl/requirements.txt
census==0.8.15
us==2.0.2
tqdm==4.49.0
pandas==1.1.2
psycopg2==2.8.6
PyYAML==5.3.1
click==7.1.2
attrdict==2.0.1
